# Improve performance of toArrayLike