--- a/benchmarks/index.js
+++ b/benchmarks/index.js
@@ -452,4 +452,10 @@ add('bitLength', {
   }
 });
 
+add('toArray', {
+  'bn.js': function () {
+    fixture.a1j.toArray();
+  }
+});
+
 start();