--- a/lib/bn.js
+++ b/lib/bn.js
@@ -540,6 +540,13 @@
     return this.toArrayLike(Array, endian, length);
   };
 
+  var allocate = function allocate (ArrayType, size) {
+    if (ArrayType.allocUnsafe) {
+      return ArrayType.allocUnsafe(size);
+    }
+    return new ArrayType(size);
+  };
+
   BN.prototype.toArrayLike = function toArrayLike (ArrayType, endian, length) {
     this._strip();
 
@@ -548,73 +555,68 @@
     assert(byteLength <= reqLength, 'byte array longer than desired length');
     assert(reqLength > 0, 'Requested array length <= 0');
 
-    var littleEndian = endian === 'le';
     var res = allocate(ArrayType, reqLength);
-
-    var i;
-    if (littleEndian) {
-      var position = 0;
-      var carry = 0;
-      for (i = 0; i < this.length; i++) {
-        var shift = (i % 4) * 2;
-        var word = (this.words[i] << shift) | carry;
-
-        res[position++] = word & 0xff;
-        res[position++] = (word >> 8) & 0xff;
-        res[position++] = (word >> 16) & 0xff;
-
-        if (shift === 6) {
-          res[position++] = (word >> 24) & 0xff;
-          carry = 0;
-        } else {
-          carry = word >>> 24;
-        }
-      }
-
-      if (position < reqLength) {
-        res[position++] = carry;
-
-        while (position < reqLength) {
-          res[position++] = 0;
-        }
-      }
-    } else {
-      // big-endian
-      var position = reqLength - 1;
-      var carry = 0;
-      for (i = 0; i < this.length; i++) {
-        var shift = (i % 4) * 2;
-        var word = (this.words[i] << shift) | carry;
-
-        res[position--] = word & 0xff;
-        res[position--] = (word >> 8) & 0xff;
-        res[position--] = (word >> 16) & 0xff;
-
-        if (shift === 6) {
-          res[position--] = (word >> 24) & 0xff;
-          carry = 0;
-        } else {
-          carry = word >>> 24;
-        }
-      }
-
-      if (position >= 0) {
-        res[position--] = carry;
-
-        while (position >= 0) {
-          res[position--] = 0;
-        }
-      }
-    }
-
+    var postfix = endian === 'le' ? 'LE' : 'BE';
+    this['_toArrayLike' + postfix](res, byteLength);
     return res;
   };
 
-  var allocate = function allocate (ArrayType, size) {
-    if (ArrayType.allocUnsafe) {
-      return ArrayType.allocUnsafe(size);
+  BN.prototype._toArrayLikeLE = function _toArrayLikeLE (res, byteLength) {
+    var position = 0;
+    var carry = 0;
+
+    for (var i = 0; i < this.length; i++) {
+      var shift = (i % 4) * 2;
+      var word = (this.words[i] << shift) | carry;
+
+      res[position++] = word & 0xff;
+      res[position++] = (word >> 8) & 0xff;
+      res[position++] = (word >> 16) & 0xff;
+
+      if (shift === 6) {
+        res[position++] = (word >> 24) & 0xff;
+        carry = 0;
+      } else {
+        carry = word >>> 24;
+      }
     }
-    return new ArrayType(size);
+
+    if (position < res.length) {
+      res[position++] = carry;
+
+      while (position < res.length) {
+        res[position++] = 0;
+      }
+    }
+  };
+
+  BN.prototype._toArrayLikeBE = function _toArrayLikeBE (res, byteLength) {
+    var position = res.length - 1;
+    var carry = 0;
+
+    for (var i = 0; i < this.length; i++) {
+      var shift = (i % 4) * 2;
+      var word = (this.words[i] << shift) | carry;
+
+      res[position--] = word & 0xff;
+      res[position--] = (word >> 8) & 0xff;
+      res[position--] = (word >> 16) & 0xff;
+
+      if (shift === 6) {
+        res[position--] = (word >> 24) & 0xff;
+        carry = 0;
+      } else {
+        carry = word >>> 24;
+      }
+    }
+
+    if (position >= 0) {
+      res[position--] = carry;
+
+      while (position >= 0) {
+        res[position--] = 0;
+      }
+    }
   };
 
   if (Math.clz32) {