--- a/lib/bn.js
+++ b/lib/bn.js
@@ -570,11 +570,17 @@
       var word = (this.words[i] << shift) | carry;
 
       res[position++] = word & 0xff;
-      res[position++] = (word >> 8) & 0xff;
-      res[position++] = (word >> 16) & 0xff;
+      if (position < res.length) {
+        res[position++] = (word >> 8) & 0xff;
+      }
+      if (position < res.length) {
+        res[position++] = (word >> 16) & 0xff;
+      }
 
       if (shift === 6) {
-        res[position++] = (word >> 24) & 0xff;
+        if (position < res.length) {
+          res[position++] = (word >> 24) & 0xff;
+        }
         carry = 0;
       } else {
         carry = word >>> 24;
@@ -599,11 +605,17 @@
       var word = (this.words[i] << shift) | carry;
 
       res[position--] = word & 0xff;
-      res[position--] = (word >> 8) & 0xff;
-      res[position--] = (word >> 16) & 0xff;
+      if (position >= 0) {
+        res[position--] = (word >> 8) & 0xff;
+      }
+      if (position >= 0) {
+        res[position--] = (word >> 16) & 0xff;
+      }
 
       if (shift === 6) {
-        res[position--] = (word >> 24) & 0xff;
+        if (position >= 0) {
+          res[position--] = (word >> 24) & 0xff;
+        }
         carry = 0;
       } else {
         carry = word >>> 24;