--- a/lib/bn.js
+++ b/lib/bn.js
@@ -541,39 +541,69 @@
   };
 
   BN.prototype.toArrayLike = function toArrayLike (ArrayType, endian, length) {
+    this._strip();
+
     var byteLength = this.byteLength();
     var reqLength = length || Math.max(1, byteLength);
     assert(byteLength <= reqLength, 'byte array longer than desired length');
     assert(reqLength > 0, 'Requested array length <= 0');
 
-    this._strip();
     var littleEndian = endian === 'le';
     var res = allocate(ArrayType, reqLength);
 
-    var b, i;
-    var q = this.clone();
-    if (!littleEndian) {
-      // Assume big-endian
-      for (i = 0; i < reqLength - byteLength; i++) {
-        res[i] = 0;
+    var i;
+    if (littleEndian) {
+      var position = 0;
+      var carry = 0;
+      for (i = 0; i < this.length; i++) {
+        var shift = (i % 4) * 2;
+        var word = (this.words[i] << shift) | carry;
+
+        res[position++] = word & 0xff;
+        res[position++] = (word >> 8) & 0xff;
+        res[position++] = (word >> 16) & 0xff;
+
+        if (shift === 6) {
+          res[position++] = (word >> 24) & 0xff;
+          carry = 0;
+        } else {
+          carry = word >>> 24;
+        }
       }
 
-      for (i = 0; !q.isZero(); i++) {
-        b = q.andln(0xff);
-        q.iushrn(8);
+      if (position < reqLength) {
+        res[position++] = carry;
 
-        res[reqLength - i - 1] = b;
+        while (position < reqLength) {
+          res[position++] = 0;
+        }
       }
     } else {
-      for (i = 0; !q.isZero(); i++) {
-        b = q.andln(0xff);
-        q.iushrn(8);
+      // big-endian
+      var position = reqLength - 1;
+      var carry = 0;
+      for (i = 0; i < this.length; i++) {
+        var shift = (i % 4) * 2;
+        var word = (this.words[i] << shift) | carry;
 
-        res[i] = b;
+        res[position--] = word & 0xff;
+        res[position--] = (word >> 8) & 0xff;
+        res[position--] = (word >> 16) & 0xff;
+
+        if (shift === 6) {
+          res[position--] = (word >> 24) & 0xff;
+          carry = 0;
+        } else {
+          carry = word >>> 24;
+        }
       }
 
-      for (; i < reqLength; i++) {
-        res[i] = 0;
+      if (position >= 0) {
+        res[position--] = carry;
+
+        while (position >= 0) {
+          res[position--] = 0;
+        }
       }
     }