--- a/lib/bn.js
+++ b/lib/bn.js
@@ -565,8 +565,7 @@
     var position = 0;
     var carry = 0;
 
-    for (var i = 0; i < this.length; i++) {
-      var shift = (i % 4) * 2;
+    for (var i = 0, shift = 0; i < this.length; i++) {
       var word = (this.words[i] << shift) | carry;
 
       res[position++] = word & 0xff;
@@ -582,8 +581,10 @@
           res[position++] = (word >> 24) & 0xff;
         }
         carry = 0;
+        shift = 0;
       } else {
         carry = word >>> 24;
+        shift += 2;
       }
     }
 
@@ -600,8 +601,7 @@
     var position = res.length - 1;
     var carry = 0;
 
-    for (var i = 0; i < this.length; i++) {
-      var shift = (i % 4) * 2;
+    for (var i = 0, shift = 0; i < this.length; i++) {
       var word = (this.words[i] << shift) | carry;
 
       res[position--] = word & 0xff;
@@ -617,8 +617,10 @@
           res[position--] = (word >> 24) & 0xff;
         }
         carry = 0;
+        shift = 0;
       } else {
         carry = word >>> 24;
+        shift += 2;
       }
     }