From 6x to 174x improvement. Was need turn off bignum, because node 12 not supported.

```bash
$ # old 768 bytes number
$ SEED=27af4987aa60091dc42b1314d340e8c0 node index.js toArray
Seed: 27af4987aa60091dc42b1314d340e8c0
Benchmarking: toArray
bn.js#toArray x 3,437 ops/sec ±7.94% (7 runs sampled)
------------------------
Fastest is bn.js#toArray
========================
$ # old 32 bytes number
$ SEED=27af4987aa60091dc42b1314d340e8c0 node index.js toArray
Seed: 27af4987aa60091dc42b1314d340e8c0
Benchmarking: toArray
bn.js#toArray x 947,728 ops/sec ±3.40% (9 runs sampled)
------------------------
Fastest is bn.js#toArray
========================
$ # new 768 bytes number
$ SEED=27af4987aa60091dc42b1314d340e8c0 node index.js toArray
Seed: 27af4987aa60091dc42b1314d340e8c0
Benchmarking: toArray
bn.js#toArray x 600,559 ops/sec ±3.99% (8 runs sampled)
------------------------
Fastest is bn.js#toArray
========================
$ # new 32 bytes number
$ SEED=27af4987aa60091dc42b1314d340e8c0 node index.js toArray
Seed: 27af4987aa60091dc42b1314d340e8c0
Benchmarking: toArray
bn.js#toArray x 6,293,522 ops/sec ±1.83% (8 runs sampled)
------------------------
Fastest is bn.js#toArray
========================
```